Patch 0768
・gcc4.3.4環境下でのコンパイル警告を一部修正 (inter.c clif.* intif.* script.c)
・anti hackerのコードを更新 (clif.c)
 ->GMでないキャラが他人に赤エモを付与出来るexploitの修正
・docの修正と追加 (client_packet.txt script_ref.txt db_ref.txt)
・packet_db.txtの修正ついでにGM右クリックメニューの全ての装備を解除機能を実装 (packet_db.txt clif.c)
・item_db.txtのtypeを本家仕様に修正 (clif.c item_db.c)
　※追加されたtypeに合わせてソースを変更しただけなので、item_db.txtの変更はまだ未着手・・・。
・PCが死亡時にコインも削除するよう変更 (pc.c)
・スクリプト関数「checkwolf()」「checkwolfmount()」を追加 (script.c script_ref.txt)
　tomoさんに感謝

+++ src/map/clif.h
@@ -299,7 +299,7 @@ void clif_guild_xy(struct map_session_data *sd);
 void clif_displaymessage(const int fd, const char* mes);
 void clif_disp_onlyself(const int fd, const char *mes);
 void clif_GMmessage(struct block_list *bl, const char* mes, size_t len, int flag);
-void clif_announce(struct block_list *bl, const char* mes, size_t len, unsigned long color, short type, short size, short align, short pos_y, int flag);
+void clif_announce(struct block_list *bl, const char* mes, size_t len, unsigned long color, int type, int size, int align, int pos_y, int flag);
 void clif_onlymessage(const char *mes, size_t len);
 void clif_heal(int fd, int type, int val);
 void clif_resurrection(struct block_list *bl, unsigned short type);
